[Feat] Add excluded files from backup

+++ scripts/backup.sh
@@ -5,6 +5,12 @@
 # This script takes care of performing a backup for pre-existing dotfiles
 # in the system
 
+BACKUPEXCLUDES=(
+  atom                    # To heavy
+  git-completion.bash     # This one doesn't change
+  git-prompt.sh           # Neither does this one
+)
+
 do_backup () {
   DOTFILES_ROOT_OLD="$HOME/.dotfiles_old"
   display_info "Creating a backup for pre-existing dotfiles..."
@@ -14,10 +20,14 @@ do_backup () {
 
   # Here we'll move every file or directory that would be overriden
   # by the symlinking step
-  for src in $( find -H "$DOTFILES_ROOT" -maxdepth 2 -name "*.symlink" -not -path '*.git*' ); do
-    overridable="$HOME/.$(basename "${src%.*}")"
-    if cp -RL "$overridable" "$DOTFILES_ROOT_OLD/" 2>/dev/null; then
-      rm "$overridable" && display_info "Moved [$src] to ~/$DOTFILES_ROOT_OLD"
+  for src in $( find -H "$DOTFILES_ROOT" -maxdepth 2 -name "*.symlink" -not -path "*.git*"  ); do
+    name=$(basename "${src%.*}")
+    # Check for file/directory that should not be backed up
+    if [[ "${BACKUPEXCLUDES[@]}" != *"$name"* ]]; then
+      overridable="$HOME/.$name"
+      if cp -RL "$overridable" "$DOTFILES_ROOT_OLD/" 2>/dev/null; then
+        rm "$overridable" && display_info "Moved [$src] to ~/$DOTFILES_ROOT_OLD"
+      fi
     fi
   done
